ref: util 함수 분리
