+++ src/main/kotlin/calculator/Util.kt
@@ -1,0 +1,9 @@
+package calculator
+
+class Util {
+    companion object {
+        fun groupByIsNumber(inputs: List<String>, bool: Boolean): List<String> {
+            return inputs.groupBy { (it.toFloatOrNull() != null) }[bool]!!
+        }
+    }
+}
